requirements.txt
Flask
Flask-Mail
python-dotenv
pydantic
sqlalchemy
datetime
python-dateutil
isodate
Flask-SQLAlchemy
gunicorn
psycopg2-binary
pytz
pandas
openpyxl
werkzeug